+++ src/components/Customers/Customers.jsx
@@ -1,0 +1,23 @@
+import photo1 from '../AdvantagesSection/img/photo1.png';
+import photo2 from '../AdvantagesSection/img/photo2.png';
+import photo3 from '../AdvantagesSection/img/photo3.png';
+
+import styles from './Customers.module.css';
+
+export default function Customers() {
+  const photos = [{ src: photo1 }, { src: photo2 }, { src: photo3 }];
+  return (
+    <div className={styles.customersBox}>
+      <ul className={styles.customersList}>
+        {photos.map((photo, idx) => (
+          <li key={idx}>
+            <img src={photo.src} />
+          </li>
+        ))}
+      </ul>
+      <p className={styles.customersText}>
+        Our <span className={styles.customersSpan}>happy</span> customers
+      </p>
+    </div>
+  );
+}
