+++ src/index.css
@@ -9,7 +9,6 @@
 
   background-color: var(--main-white-color);
   color: var(--main-black-gray-color);
-
   
 }
